chore: code review, requests
